PROBLEM: lot's of repetition in Kubernetes YAML files
SOLUTION: support partials

The implemented partials system is very simplistic, although it looks
a bit like Rails templating. There's only one function supported, called
"partial", which takes two arguments: the name of the partial and an
argument hash. Partials are expected to be placed in a "partials"
subdirectory of the template directory and must have a '.yaml.erb'
suffix. Keys in the arguments hash are converted into local variables
for the partial. They are also accessible via the argument hash which is
passed in as variable 'locals'.